bundle/spec.go: manually use preserve=all option instead of archive
The archive option won't follow symlinks, so if the runrootless-proot file inside ~/.runrootless/runrootless-proot is a symlink we end up copying a symlink into the container which won't work cause the container won't be able to follow the symlink outside of the container. Instead just copy all permissions, which is what we really want to do here instead.